+++ src/pages/Settings/display/AppLocker.tsx
